# server/openapi.yaml
openapi: 3.1.0
info:
  title: Voxels public read API
  version: '1.0.0'
  description: |
    The public Voxels API. Most of this is unauthenticated GET.

    Write routes for collection create / deploy / mint writeback are documented
    here. Sign-in, admin, moderation, livekit, radio, metrics, guest passes and
    the internal /grid/* routes are deliberately not described.

    Most handlers answer with an envelope: `{"success": true, "<field>": ...}`,
    where `<field>` is plural for a list and singular for one record. A failed
    lookup answers `{"success": false}`, usually with status 400 and sometimes
    with 200, so check the flag rather than the status.
servers:
  - url: https://www.voxels.com
...
paths:
  /federation/edit/{id}:
    get:
      tags: [federation]
      summary: Review a retained public signed edit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, pattern: '^0x[a-f0-9]{64}$' }
      responses:
        '200':
          description: The complete signed edit or permission change
        '404':
          description: Public edit not found
...
  /federation/parcel/{id}:
    get:
      tags: [federation]
      summary: Read the public parcel genesis hash and latest causal revision
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Base hash, causal clock, parent, active parcel permission revision, preserved edit IDs and pending review IDs
        '400':
          description: Invalid or non-public parcel
  /federation/edits:
    post:
      tags: [federation]
      summary: Verify and durably apply a wallet-signed public build edit
      description: Version 2 requires a parcel owner, manager or builder wallet signature bound to a permission revision. Hosts independently verify archived authority and signed permission changes. Conflicting manager branches pause writes until the parcel owner resolves them. Superseded-authority edits remain pending unless explicitly preserved or approved. Version 1 remains restricted to the network owner for compatibility. Legacy account tokens do not authorize edits. See selfhost/node/README.md and common/federation/edit.ts for canonical signing, supported patches and bounds. Requests are limited to 2 MiB; signed envelopes to 1,900,000 UTF-8 bytes.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Accepted edit identifier; duplicate submissions are idempotent
        '400':
          description: Invalid proof, missing parent, wrong snapshot or unsupported build
  /federation/batch:
    post:
      tags: [federation]
      summary: Receive up to 16 independently signed edits
      description: Events are verified in order. A failure may follow earlier committed events; retrying is safe because accepted IDs are idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  maxItems: 16
                  items: { type: object }
      responses:
        '200':
          description: Accepted identifiers
        '400':
          description: Invalid batch or edit
  /federation/events:
    get:
      tags: [federation]
      summary: Read a bounded feed of committed public edits
      parameters:
        - name: after
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: World identifier, up to 16 sequence/event entries and next cursor
        '400':
          description: Invalid cursor
  /federation/presence:
    get:
      tags: [federation]
      summary: Read anonymous, short-lived host-signed player poses
      responses:
        '200':
          description: World identifier and at most 16 valid host snapshots
    post:
      tags: [federation]
      summary: Relay host-signed player poses
      description: Only world, node, timestamp and anonymous pose fields are accepted. Validity lasts 15 seconds. Host signatures do not prove human identity; unverified wallet/name claims are not forwarded.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshots]
              properties:
                snapshots:
                  type: array
                  maxItems: 16
                  items: { type: object }
      responses:
        '200':
          description: Envelope processed; invalid or expired snapshots are ignored
        '400':
          description: Invalid envelope
...